Move theme default to index and remove babelrc
